Modernize to Node 18-22 + TensorFlow.js 4.x
Bump the 2020-era toolchain to current: @tensorflow/tfjs-core 1.7 -> 4.22,
TypeScript 3.8 -> 5.9, Rollup 2 -> 4, canvas 2.6 -> 3, Node <=12 -> 18/20/22.

Model code and shipped weights are unchanged. Source edits are mechanical
rewrites of tfjs APIs removed since 1.7 (chained Tensor methods -> functional
tf.* ops: toFloat/toInt -> cast, asXD -> reshape, div/mul/sub/mean/expandDims/
slice/reshape), plus strict-TS fixes exposed by dropping the removed
suppressImplicitAnyIndexErrors option.

Build:
- rollup.config.js -> rollup.config.mjs on @rollup/* plugins
- tsconfig target/lib es5 -> es2018
- .travis.yml/karma.conf.js/jasmine-node.js removed (dead deps);
  CI is now .github/workflows/ci.yml on Node 18/20/22

Validation: scripts/smoke.js runs on @tensorflow/tfjs-node and checks the
migrated build against the repo's frozen fixtures -- faceRecognitionNet
descriptor within euclid 0.036 of test/data/faceDescriptor1.json (tol 0.1),
tinyFaceDetector boxes within 2px of the reference, full high-level pipeline
well-formed. `npm test` = build + smoke.

Full Karma/Jasmine suite port is pending; see MIGRATION.md.

+++ src/mtcnn/Mtcnn.ts
@@ -57,7 +57,7 @@ export class Mtcnn extends NeuralNetwork<NetParams> {
 
     const imgTensor = tf.tidy(() =>
       bgrToRgbTensor(
-        tf.expandDims(tf.browser.fromPixels(inputCanvas)).toFloat() as tf.Tensor4D
+        tf.cast(tf.expandDims(tf.browser.fromPixels(inputCanvas)), 'float32') as tf.Tensor4D
       )
     )
